Aktarian Starship -  A blender project
http://www.starship-whisperer.com/

I started the website, using Notepad++ in November 2014 with every intention of simply writing a book via the site. However, I got sidetracked by wishing to illustrate the ships and had to learn how to use Blender in order to do this. This project was started in February, at the same time as I began blender.

Katiya and Gemima are half human, half starship, their Mother's hybrid half is an unknown race several universes over.